Remove `AbstractShow*` classes
There's only a single implementation so the abstract classes can be
removed.
